Rename package - persistance -> persistence (typo)
