+++ ws/src/main/scala/de/thm/ii/fbs/services/persistence/CourseRegistrationService.scala
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.services.persistance
+package de.thm.ii.fbs.services.persistence
 
 import java.sql.ResultSet
